+++ HonorAmongThieves/wwwroot/js/heist.js
@@ -278,50 +278,3 @@ connection.on("EndGame_Broadcast", function (year, leaderboarddata) {
     sessionStorage.removeItem("username");
     sessionStorage.removeItem("roomid");
 });
-
-connection.start().catch(function (err) {
-    return console.error(err.toString());
-});
-
-connection.on("FreshConnection", function () {
-    var sessionUserName = sessionStorage.getItem("username");
-    var sessionRoomId = sessionStorage.getItem("roomid");
-    //reconnecting = false;
-
-    if (sessionUserName != null && sessionRoomId != null) {
-        // Resume the session
-        userName = sessionUserName;
-        roomId = sessionRoomId;
-
-        connection.invoke("ResumeSession", roomId, userName).catch(function (err) {
-            return console.error(err.toString());
-        });
-    }
-});
-
-connection.on("ClearState", function () {
-    sessionStorage.removeItem("username");
-    sessionStorage.removeItem("roomid");
-})
-
-var conditionalReload = function () {
-    var sessionUserName = sessionStorage.getItem("username");
-    var sessionRoomId = sessionStorage.getItem("roomid");
-
-    if (sessionUserName != null && sessionRoomId != null) {
-        //reconnecting = true;
-        var elements = document.getElementsByClassName("state");
-        for (var i = 0; i < elements.length; i++) {
-            elements[i].style.display = "none";
-        }
-
-        document.getElementById("pageName").textContent = "RECONNECTING...";
-        //setInterval(function () {
-        //    if (reconnecting === true) {
-        //        window.location.reload()
-        //    }
-        //}, 10000)
-    }
-}
-
-window.onload = conditionalReload;
